fix: create a wrapper for the server component and modified it to accept params as Promise and await it
